fix: correcion de estructura del proyecto. estructura correcta siguiendo el estandar de firebase functions
