+++ functions/package.json
@@ -15,10 +15,14 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "dotenv": "^16.5.0",
+    "express": "^5.1.0",
     "firebase-admin": "^12.6.0",
-    "firebase-functions": "^6.0.1"
+    "firebase-functions": "^6.0.1",
+    "pg": "^8.16.0"
   },
   "devDependencies": {
+    "@types/pg": "^8.15.2",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
     "eslint": "^8.9.0",
