--- package.json
@@ -1,36 +1,0 @@
-{
-  "name": "microservice-chatbotia",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "ts-node src/index.ts"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/GestionCursos/microservice-ChatBotIA.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
-  "bugs": {
-    "url": "https://github.com/GestionCursos/microservice-ChatBotIA/issues"
-  },
-  "homepage": "https://github.com/GestionCursos/microservice-ChatBotIA#readme",
-  "dependencies": {
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
-    "pg": "^8.16.0"
-  },
-  "devDependencies": {
-    "@types/express": "^5.0.2",
-    "@types/node": "^22.15.18",
-    "@types/pg": "^8.15.2",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.8.3"
-  }
-}
